clean up the interface between ncc.c, gen.c and out.c
out.c was changes and simplified so that it could be easily
implemented for other executable formats like a.out.  The
interface between ncc and gen has been simplified a lot too.
Now relocations and symbols are indicated with a name instead
of symbol table offset, so even for function static variables
and string literals a unique local name is created.